# munayfund-api2/docs/swagger.yaml
basePath: /api/v1
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  munayfund-api2_internal_core_domain.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  munayfund-api2_internal_core_domain.Multimedia:
    properties:
      type:
        type: string
      url:
        type: string
    type: object
  munayfund-api2_internal_core_domain.Project:
    properties:
      advances:
        items:
          $ref: '#/definitions/munayfund-api2_internal_core_domain.ProjectAdvance'
        type: array
      description:
        type: string
      id:
        type: string
      multimedia:
        items:
          $ref: '#/definitions/munayfund-api2_internal_core_domain.Multimedia'
        type: array
      title:
        type: string
      userId:
        type: string
    type: object
  munayfund-api2_internal_core_domain.ProjectAdvance:
    properties:
      date:
        type: string
      details:
        type: string
      media:
        items:
          $ref: '#/definitions/munayfund-api2_internal_core_domain.Multimedia'
        type: array
    type: object
  munayfund-api2_internal_core_domain.User:
    properties:
      email:
        type: string
      id:
        type: string
      password:
        type: string
      username:
        type: string
      wallet_id:
        type: string
    type: object
...
host: localhost:8085
info:
  contact:
    email: todefine@github.com
    name: KaniKrafters
    url: https://todefine.github.com
  description: This API is in charge of leveraging some functionalities of munayFund
    a next gen crowdfunding solution.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: Swagger MunayFund API WEB 2
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Este endpoint permite a un usuario autenticarse, proporcionando
        el email y la contraseña.
      operationId: loginUser
      parameters:
      - description: Credenciales de inicio de sesión del usuario
        in: body
        name: loginUser
        required: true
        schema:
          $ref: '#/definitions/munayfund-api2_internal_core_domain.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: Proyecto actualizado exitosamente
          schema:
            $ref: '#/definitions/munayfund-api2_internal_core_domain.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Realiza la autenticación de un usuario.
      tags:
      - users
...
  /signup:
    post:
      consumes:
      - application/json
      description: Este endpoint permite a un usuario registrarse proporcionando la
        información requerida, incluyendo email y contraseña.
      operationId: signUpUser
      parameters:
      - description: Información del nuevo usuario
        in: body
        name: signUpUser
        required: true
        schema:
          $ref: '#/definitions/munayfund-api2_internal_core_domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: Token de autenticación generado
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Registra a un nuevo usuario en el sistema.
      tags:
      - users
  /user/{id}:
    delete:
      description: Este endpoint permite a un usuario eliminar su cuenta proporcionando
        su ID.
      operationId: deleteUser
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del usuario a eliminar
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Respuesta de éxito al eliminar el usuario
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Elimina un usuario existente.
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Este endpoint permite a un usuario actualizar su información proporcionando
        los datos requeridos.
      operationId: updateUser
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del usuario a actualizar
        in: path
        name: id
        required: true
        type: string
      - description: Información actualizada del usuario
        in: body
        name: updateUser
        required: true
        schema:
          $ref: '#/definitions/munayfund-api2_internal_core_domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: Usuario actualizado
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Actualiza la información del usuario.
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"